+++ src/Common/DB/Crud/CrudTransaction.cs
@@ -3,9 +3,7 @@ namespace Common.DB.Crud;
 using System;
 using System.Threading.Tasks;
 
-public class CrudTransaction(
-    CrudEntry[] crud,
-    Func<string?, Task> complete, int? transactionId = null) : CrudBatch(crud, false, complete)
+public class CrudTransaction(CrudEntry[] crud, Func<string?, Task> complete, long? transactionId = null) : CrudBatch(crud, false, complete)
 {
-    public int? TransactionId { get; private set; } = transactionId;
+    public long? TransactionId { get; private set; } = transactionId;
 }
